feat(home): update and fix issue home
